Add files via upload
Create a BPMN diagram for the following requirements.
Helpdesk
The process of handling a tech support request starts with entering a new ticket by the customer or helpdesk employee. At this point parallel work on technical and business sides of the request starts.
Business side: Customer service department employee verifies, whether the request is covered by SLA. If it is, no extra activity is required. If it is not, the customer service department sends an invoice to the customer.
Technical side:On the helpdesk side, one of the employees of the helpdesk department is assigned to the ticket. He then consults the knowledge base in search of similar problems. If the solution to the problem can be found in the knowledge base, it is used to solve the problem. If it is not - the employee solves the problem on his own and records the solution in the knowledge base. 
The solution is then sent to the customer.